+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Add --only-safe option to pick command (default: `true`).
+
 ## 0.5.0 - 2022-05-16
 
 ### Changed